live_dashboard/requirements.txt
dash
dash_bootstrap_components
psycopg2-binary
sqlalchemy
pandas
python-dotenv
plotly
nbformat==5.1.2